gh-130: Renaming user files to player ones
