+++ frontend/src/app/components/Leaderboard/Leaderboard.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 import { ListCon } from '../../../styles/blocks'
 import { LeaderboardRow } from './LeaderboardRow'
 import { LeaderboardFilters } from './LeaderboardFilters'
-import { getTopUsers } from '../../modules/users/users-selectors'
+import { getTopUsers } from '../../modules/players/players-selectors'
 import { LeaderboardActions } from '../../modules/leaderboard/leaderboard-actions'
 import * as Filters from '../../const/leaderboard-filters'
