+++ frontend/src/app/modules/players/players-actions.js
@@ -5,5 +5,5 @@ export const UsersActions = createActions({
   addUser: ['user'],
   usersLoaded: ['users'],
 }, {
-  prefix: 'foosball-rating/users/',
+  prefix: 'foosball-rating/players/',
 })
