+++ frontend/src/app/modules/players/players-selectors.js
