+++ frontend/src/app/pages/AddPlayer/index.js
@@ -1,4 +1,4 @@
 import React from 'react'
-import { AddUser } from '../../components/AddUser/AddUser'
+import { AddUser } from '../../components/AddPlayer/AddPlayer'
 
 export const AddUserPage = () => <AddUser />
