+++ frontend/src/app/pages/Profile/index.js
@@ -7,7 +7,7 @@ import {
 } from './../../../styles/blocks'
 import { ProfileBattleHistory } from './../../components/ProfileBattleHistory'
 import { ProfileRatingGraph } from '../../components/ProfileRatingGraph'
-import { getUser } from '../../modules/users/users-selectors'
+import { getUser } from '../../modules/players/players-selectors'
 import { getStatisticsForUser } from '../../modules/matches/matches-selectors'
 
 class ProfileComponent extends Component {
